+++ ChangeLog.md
@@ -11,6 +11,7 @@
 * --force-dirty flag: Force treating all local packages as having dirty files (useful for cases where stack can't detect a file change)
 * GHC error messages: display file paths as absolute instead of relative for better editor integration
 * Add the `--copy-bins` option [#569](https://github.com/commercialhaskell/stack/issues/569)
+* Give warnings on unexpected config keys [#48](https://github.com/commercialhaskell/stack/issues/48)
 
 Bug fixes: